refactor(shares): extract formatRelative to shared util
